Update PR description formatting.
Previously, the commit description was dropped after the stack list and was in a pre-formatted block, which removes any markdown formatting and other niceties of description commit bodies.

This instead pulls the commit body to the top, and adds a horizontal divider, after which the SPR stack and other trailer info is added. This keeps the PR description more in line with the commit message, while still including the Stack info.

commit-id:e0bb6ce4